Настройка swaggerUI
