+++ order/src/main/java/org/dclou/example/demogpb/order/config/WebMvcConfiguration.java
@@ -1,0 +1,29 @@
+package org.dclou.example.demogpb.order.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+/**
+ * Created by Александр Подоплелов on 06.06.2017.
+ */
+@Configuration
+public class WebMvcConfiguration extends WebMvcConfigurerAdapter {
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addRedirectViewController("/documentation/v2/api-docs", "/v2/api-docs");
+        registry.addRedirectViewController("/documentation/configuration/ui", "/configuration/ui");
+        registry.addRedirectViewController("/documentation/configuration/security", "/configuration/security");
+        registry.addRedirectViewController("/documentation/swagger-resources", "/swagger-resources");
+        registry.addRedirectViewController("/documentation", "/documentation/swagger-ui.html");
+        registry.addRedirectViewController("/documentation/", "/documentation/swagger-ui.html");
+
+    }
+
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/documentation/**").addResourceLocations("classpath:/META-INF/resources/");
+    }
+
+}
